Step 3.2 - Added LocationResolver, so the resolving of the classpath can be dispatched according to Java version
